refactor(features): modularize person and transaction modals
This refactor moves modal generation logic into dedicated feature modules to improve separation of concerns.

- Move person modals (`showPersonModal`, `editPerson`) to `features/persons/person-modals.js`.
- Move transaction modals (`showTransactionModal`, etc.) to `features/transactions/transaction-modals.js`.
- Update `core/main.js` to initialize the new modules.
- Update `ui/fab.js` and `ui/renderer.js` to import modal functions directly from their new locations.
- Remove placeholder modal functions from `core/state.js`.
