+++ core/main.js
@@ -1,11 +1,10 @@
 // core/main.js
 
 import { db } from '../db.js';
-import { generatePersonId, pickContact } from '../contact-helper.js';
 import { app } from './state.js';
 
 // Import UI Modules
-import { initModal, showModal, closeModal } from '../ui/modal.js';
+import { initModal } from '../ui/modal.js';
 import { initFab } from '../ui/fab.js';
 import { initNavigation } from '../ui/navigation.js';
 import { showAlert, showConfirm } from '../ui/notifications.js';
@@ -13,6 +12,8 @@ import { initRenderer, render } from '../ui/renderer.js';
 
 // Import Feature Modules
 import { initActions, deleteTransaction, deletePayment } from '../features/actions.js';
+import { initPersonModals } from '../features/persons/person-modals.js';
+import { initTransactionModals, showEditTransactionModal, showPaymentModal, showTransactionDetails } from '../features/transactions/transaction-modals.js';
 
 
 // =================================================================================================
@@ -30,7 +31,6 @@ export async function init() {
     initRenderer({
         calculateBalance,
         handleTransactionAction,
-        editPerson,
     });
 
     // Initialize feature modules
@@ -40,10 +40,22 @@ export async function init() {
         calculateBalance,
     });
 
+    initPersonModals({
+        loadData,
+        render,
+    });
+
+    initTransactionModals({
+        calculateBalance,
+        deletePayment,
+        loadData,
+        render,
+    });
+
 
     // Initialize UI modules, passing the app object for context
     initModal({ render });
-    initFab({ ...app, showPersonModal, showTransactionModal }); // Pass app state and modal functions
+    initFab();
     initNavigation({ render });
 
     setupEventListeners();
@@ -117,247 +129,6 @@ function handleTransactionAction(e) {
             deleteTransaction(id);
             break;
     }
-}
-
-/**
- * Updates a person's details in the database.
- * @param {string} personId - The ID of the person to edit.
- */
-function editPerson(personId) {
-    const person = app.persons.find(p => p.id === personId);
-
-    showModal('Edit Person', `
-    <form id="editPersonForm">
-      <div class="form-group"><label class="label">First Name</label><input type="text" name="firstName" class="input" value="${person.firstName}" required></div>
-      <div class="form-group"><label class="label">Last Name</label><input type="text" name="lastName" class="input" value="${person.lastName || ''}"></div>
-      <div class="form-group"><label class="label">Phone Number</label><input type="tel" name="phone" class="input" value="${person.phone}" required></div>
-      <button type="submit" class="btn w-full">Update</button>
-    </form>
-  `);
-
-    document.getElementById('editPersonForm').addEventListener('submit', async (e) => {
-        e.preventDefault();
-        const formData = new FormData(e.target);
-
-        person.firstName = formData.get('firstName');
-        person.lastName = formData.get('lastName');
-        person.phone = formData.get('phone');
-
-        const newId = await generatePersonId(person.firstName, person.phone);
-
-        if (newId !== person.id) {
-            app.transactions.forEach(t => {
-                if (t.personId === person.id) t.personId = newId;
-            });
-            await db.delete('persons', person.id);
-            for (const t of app.transactions.filter(t => t.personId === newId)) {
-                await db.put('transactions', t);
-            }
-            person.id = newId;
-        }
-
-        await db.put('persons', person);
-        await loadData();
-        closeModal();
-        render();
-    });
-};
-
-
-// =================================================================================================
-// MODAL DIALOGS
-// =================================================================================================
-
-/**
- * Shows the modal for adding a new person.
- */
-function showPersonModal() {
-    showModal('Add Person', `
-    <form id="personForm">
-      <div class="form-group"><label class="label">First Name</label><input type="text" name="firstName" class="input" required></div>
-      <div class="form-group"><label class="label">Last Name</label><input type="text" name="lastName" class="input"></div>
-      <div class="form-group"><label class="label">Phone Number</label><input type="tel" name="phone" class="input" required></div>
-      <div class="flex gap-2">
-        <button type="submit" class="btn flex-1">Save</button>
-        <button type="button" id="pickContactBtn" class="btn btn-secondary flex-1">🕵️ Pick Contact</button>
-      </div>
-    </form>
-  `);
-
-    document.getElementById('pickContactBtn').addEventListener('click', async () => {
-        try {
-            const contact = await pickContact();
-            if (contact) {
-                document.querySelector('[name="firstName"]').value = contact.firstName;
-                document.querySelector('[name="lastName"]').value = contact.lastName;
-                document.querySelector('[name="phone"]').value = contact.phone;
-            }
-        } catch (err) {
-            showAlert('Could not pick contact: ' + err.message);
-        }
-    });
-
-    document.getElementById('personForm').addEventListener('submit', async (e) => {
-        e.preventDefault();
-        const formData = new FormData(e.target);
-        const person = {
-            id: await generatePersonId(formData.get('firstName'), formData.get('phone')),
-            firstName: formData.get('firstName'),
-            lastName: formData.get('lastName'),
-            phone: formData.get('phone')
-        };
-        await db.put('persons', person);
-        await loadData();
-        closeModal();
-        render();
-    });
-}
-
-/**
- * Shows the modal for adding a new transaction.
- * @param {string} type - The type of transaction ('IOU' or 'UOM').
- */
-function showTransactionModal(type) {
-    const personOptions = app.persons.map(p => `<option value="${p.id}">${p.firstName} ${p.lastName}</option>`).join('');
-    showModal(`Add ${type}`, `
-    <form id="transactionForm">
-      <div class="form-group"><label class="label">Person</label><select name="personId" class="select" required><option value="">Select person...</option>${personOptions}</select></div>
-      <div class="form-group"><label class="label">Amount</label><input type="number" step="0.01" name="amount" class="input" placeholder="0.00" required></div>
-      <div class="form-group"><label class="label">Description</label><input type="text" name="description" class="input" required></div>
-      <div class="form-group"><label class="label">Date</label><input type="date" name="date" class="input" required value="${new Date().toISOString().split('T')[0]}"></div>
-      <div class="form-group"><label class="label">Due Date (optional)</label><input type="date" name="dueDate" class="input"></div>
-      <button type="submit" class="btn w-full">Save ${type}</button>
-    </form>
-  `);
-
-    document.getElementById('transactionForm').addEventListener('submit', async (e) => {
-        e.preventDefault();
-        const formData = new FormData(e.target);
-        const transaction = {
-            id: Date.now().toString(),
-            personId: formData.get('personId'),
-            type: type.toUpperCase(),
-            description: formData.get('description'),
-            amount: Math.round(parseFloat(formData.get('amount')) * 100),
-            date: formData.get('date'),
-            dueDate: formData.get('dueDate') || null,
-            status: 'pending',
-            payments: []
-        };
-        await db.put('transactions', transaction);
-        await loadData();
-        closeModal();
-        render();
-    });
-}
-
-/**
- * Shows the modal for adding a payment to a transaction.
- * @param {object} transaction - The transaction to add a payment to.
- */
-function showPaymentModal(transaction) {
-    const person = app.persons.find(p => p.id === transaction.personId);
-    const balance = calculateBalance(transaction);
-
-    showModal('Record Payment', `
-    <form id="paymentForm">
-      <div class="mb-2"><strong>${person?.firstName} ${person?.lastName}</strong><br><span class="text-sm text-gray">Balance: ${(balance / 100).toFixed(2)}</span></div>
-      <div class="form-group"><label class="label">Amount</label><input type="number" step="0.01" name="amount" class="input" placeholder="0.00" required max="${balance / 100}"></div>
-      <div class="form-group"><label class="label">Payment Date</label><input type="date" name="paymentDate" class="input" required value="${new Date().toISOString().split('T')[0]}"></div>
-      <div class="form-group"><label class="label">Note (optional)</label><input type="text" name="note" class="input" placeholder="Payment note"></div>
-      <button type="submit" class="btn w-full">Record Payment</button>
-    </form>
-  `);
-
-    document.getElementById('paymentForm').addEventListener('submit', async (e) => {
-        e.preventDefault();
-        const formData = new FormData(e.target);
-        const payment = {
-            id: Date.now().toString(),
-            transactionId: transaction.id,
-            amount: Math.round(parseFloat(formData.get('amount')) * 100),
-            date: formData.get('paymentDate') + 'T12:00:00.000Z',
-            note: formData.get('note')
-        };
-        transaction.payments = transaction.payments || [];
-        transaction.payments.push(payment);
-        if (calculateBalance(transaction) === 0) transaction.status = 'paid';
-        await db.put('transactions', transaction);
-        await loadData();
-        closeModal();
-        render();
-    });
-}
-
-/**
- * Shows the modal with detailed information about a transaction.
- * @param {object} transaction - The transaction to show details for.
- */
-function showTransactionDetails(transaction) {
-    const person = app.persons.find(p => p.id === transaction.personId);
-    const balance = calculateBalance(transaction);
-    const paymentsHtml = transaction.payments?.map(p => `
-    <div class="list-item flex-between">
-        <div>
-          <div class="text-sm">${(p.amount / 100).toFixed(2)}</div>
-          <div class="text-xs text-gray">${new Date(p.date).toLocaleDateString()}</div>
-          ${p.note ? `<div class="text-xs text-gray">${p.note}</div>` : ''}
-        </div>
-        <button class="btn-icon text-red" data-action="delete-payment" data-payment-id="${p.id}" data-transaction-id="${transaction.id}">×</button>
-    </div>`
-    ).join('') || '<p class="text-sm text-gray">No payments yet</p>';
-
-    showModal('Transaction Details', `
-    <div class="mb-4"><strong>${person?.firstName} ${person?.lastName}</strong><br><span class="text-sm text-gray">${transaction.description}</span></div>
-    <div class="mb-4">
-      <div class="flex-between mb-2"><span>Original Amount:</span><span>${(transaction.amount / 100).toFixed(2)}</span></div>
-      <div class="flex-between mb-2"><span>Current Balance:</span><span class="font-bold">${(balance / 100).toFixed(2)}</span></div>
-      <div class="flex-between"><span>Status:</span><span class="${transaction.status === 'paid' ? 'text-green' : ''}">${transaction.status}</span></div>
-    </div>
-    <h3 class="font-bold mb-2">Payment History</h3><div class="list" id="paymentListContainer">${paymentsHtml}</div>
-  `);
-
-    // Attach event listener to the container for payment deletion
-    document.getElementById('paymentListContainer').addEventListener('click', (e) => {
-        if (e.target.dataset.action === 'delete-payment') {
-            const paymentId = e.target.dataset.paymentId;
-            const transactionId = e.target.dataset.transactionId;
-            deletePayment(transactionId, paymentId);
-        }
-    });
-}
-
-
-/**
- * Shows the modal for editing an existing transaction.
- * @param {object} transaction - The transaction to edit.
- */
-function showEditTransactionModal(transaction) {
-    const personOptions = app.persons.map(p => `<option value="${p.id}" ${p.id === transaction.personId ? 'selected' : ''}>${p.firstName} ${p.lastName}</option>`).join('');
-    showModal(`Edit ${transaction.type}`, `
-    <form id="editTransactionForm">
-      <div class="form-group"><label class="label">Person</label><select name="personId" class="select" required>${personOptions}</select></div>
-      <div class="form-group"><label class="label">Amount</label><input type="number" step="0.01" name="amount" class="input" value="${(transaction.amount / 100).toFixed(2)}" required></div>
-      <div class="form-group"><label class="label">Description</label><input type="text" name="description" class="input" value="${transaction.description}" required></div>
-      <div class="form-group"><label class="label">Date</label><input type="date" name="date" class="input" value="${transaction.date}" required></div>
-      <div class="form-group"><label class="label">Due Date (optional)</label><input type="date" name="dueDate" class="input" value="${transaction.dueDate || ''}"></div>
-      <button type="submit" class="btn w-full">Update ${transaction.type}</button>
-    </form>
-  `);
-
-    document.getElementById('editTransactionForm').addEventListener('submit', async (e) => {
-        e.preventDefault();
-        const formData = new FormData(e.target);
-        transaction.personId = formData.get('personId');
-        transaction.description = formData.get('description');
-        transaction.amount = Math.round(parseFloat(formData.get('amount')) * 100);
-        transaction.date = formData.get('date');
-        transaction.dueDate = formData.get('dueDate') || null;
-        await db.put('transactions', transaction);
-        await loadData();
-        closeModal();
-        render();
-    });
 }
 
 // =================================================================================================
