+++ core/state.js
@@ -1,10 +1,5 @@
 // core/state.js
 
-// We will define these later, but the state object needs placeholders
-function showPersonModal() {}
-function showTransactionModal() {}
-
-
 // Global app state
 export const app = {
     version: '1.0.2',
@@ -13,6 +8,4 @@ export const app = {
     currentView: 'iou',
     currentTransaction: null,
     chart: null,
-    showPersonModal,
-    showTransactionModal,
 };
